Refactor: Tables and helper functions
